feat: added form validation
- splited Header.tsx into {Header.tsx, LogIn.tsx, LogOut.tsx, SignUp.tsx}.
