+++ src/components/Header/index.ts
@@ -1,0 +1,4 @@
+export * from "./Header";
+export * from "./LogIn";
+export * from "./SignUp";
+export * from "./LogOut";
